pyproject.toml
[project]
name = "clinicos-demo-backend"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "django>=6.0.2",
    "django-cors-headers>=4.9.0",
    "django-environ>=0.12.0",
    "django-filter>=25.2",
    "djangorestframework>=3.16.1",
    "gunicorn>=25.1.0",
    "psycopg[binary]>=3.3.2",
    "whitenoise>=6.11.0",
]
